+++ src/client/java/tempeststudios/inventorysort/InventorySearchClient.java
@@ -1,6 +1,8 @@
 package tempeststudios.inventorysort;
 
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
+import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import org.slf4j.Logger;
 import tempeststudios.inventorysort.core.InventorySortCore;
@@ -13,5 +15,10 @@ public class InventorySearchClient implements ClientModInitializer {
         LOGGER.info("Initializing Inventory Search Mod");
         InventorySearchFeature.initialize();
         ClientTickEvents.END_CLIENT_TICK.register(InventorySearchFeature::sampleInventory);
+        ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> {
+            dispatcher.register(ClientCommandManager.literal("inventorysearch")
+                    .then(WorldProfileCommands.build()));
+            LOGGER.info("Registered inventory search commands");
+        });
     }
 }